Separate notebook kernel and api test.
